FIXED: Anonymous Users Can Again See Site

+++ src/App/Firebase/FirebaseStorageService.ts
@@ -13,6 +13,7 @@ export const createImageInFirebase = (fileName: string, image: Blob): UploadTask
   if (!trimmedFileName) throw new Error('file name is empty');
   if (!image || image.size === 0) throw new Error('image is empty');
   const currentUser = readCurrentUser();
+  if (!currentUser) throw new Error('No current user');
   const storage = getStorage();
   const fileRef = ref(storage, `${currentUser.uid}/images/${trimmedFileName}`);
   return uploadBytesResumable(fileRef, image);
@@ -22,6 +23,7 @@ export const deleteImageInFirebase = (fileName: string): Promise<void> => {
   const trimmedFileName = fileName.trim();
   if (!trimmedFileName) throw new Error('file name is empty');
   const currentUser = readCurrentUser();
+  if (!currentUser) throw new Error('No current user');
   const storage = getStorage();
   const fileRef = ref(storage, `${currentUser.uid}/images/${trimmedFileName}`);
   return deleteObject(fileRef);
